speech_to_text/python/real_time_transcription/proxy/server/requirements.txt
python-dotenv==1.1.0
websockets==14.2
